BackService/requirements.txt
Django==3.2.9
mysqlclient==2.0.3
djangorestframework==3.12.4
django-cors-headers==3.10.0
colorlog
asgiref
pytz
sqlparse
numpy
psutil
opencv-python
Pillow